Add files via upload
The initial cluster centres were considered to be: µ1 = (6.2, 3.2) (red), µ2 = (6.6, 3.7) (green), µ3 = (6.5, 3.0) (blue)

Image Quantization: 

Given the image of Baboon, we needed to apply image quantization for a different number of means.
We use simple K mean clustering to image img1. The image was treated as a matrix, all pixels of the image was assigned to the all randomly generated centroids at first. We computed the distance of all pixels from the corresponding means and the least distance was selected. The pixel is added to the cluster of the center, with which, the least distance was computed. Once all the values were computed, The average of all pixel values belonging to the cluster was computed. The average value is assigned to the centroid and hence the clusters were formed. All these updated pixel values are assigned to new_image. We then take an array of random colours, pick colours needed according to the number of clusters and assign these colours to the pixels of new_image we computed in the previous part.

Image quantization is a strong tool for inferring more from a minimum amount of data. It is also used as an image segmentation and information transfer.

In this task, 

We have also calculated the loss generated by the image quantization.